update processor impl to call process() on each front duration.
This updates the processor behaviour such that process() will be called with the
minimum front duration of all input streams.  This simplifies implementations
because they do not have to keep track of the consumed tuples on each input.
Once the front tuple (with the given duration) is consumed, derived
implementations can return the duration and will be called again if tuples are
available.