build: use HINTS instead of PATHS
In some circumstances, find_path accepts a directory with the given name
as a valid search result.  With PATHS, the SystemC home directory will
be appended to the list of searched paths.  It will therefore be
searched last, after trying out an internal combination of default paths
which might already trigger a positive search
result (e.g. /opt/systemc).

Using HINTS instead of PATHS solves the problem by prepending the custom
path to the search list.

This fixes #7.